Shifted work to ShooterCharacter files after inheritance issues
Moved all work from CharAbilities and CharAbilitiesMovementComponent to ShooterCharacter and ShooterCharacterMovement after running into problems trying to include CharAbilities's functionality into the Player Pawn. Was hoping to use the functionality from ShooterCharacter AND CharAbilities, but that proved difficult. Instead, the segments that are changed in ShooterCharacter and ShooterCharacterMovement are now designated by comment sections.