Initial Networking Progress
There are a couple of tweaks that need to be made - I am unable to use the jetpack in multiplayer, and the teleport works when running as a Listen Server but not when ran as a Client.

That said, the initial networking elements are implemented. As I continue to research the network side of things I will continue to debug and fix the current issues.